+++ src/tx.rs
@@ -1,4 +1,4 @@
-use std::collections::{BTreeMap, BTreeSet, HashMap};
+use std::collections::{BTreeMap, HashMap};
 
 use anyhow::Result;
 use edn::kw;
@@ -6,12 +6,9 @@ use edn::symbols::Keyword;
 
 use crate::codec::{self, encode_datatype, encode_i64_bytes, Encode};
 use crate::indexer::vae_key_to_parts;
-use crate::metadata::PartitionMap;
 use crate::ops::{DataType, Datom, DatomOp, Entid, EntityRef, TxOp};
-use crate::partition::{dominant_partition, DB_PARTITION, DEFAULT_PARTITION, USER_PARTITION};
 use crate::schema::{Schema, Unique, ValueType};
 use crate::slate::DEFAULT_SCAN_OPTIONS;
-use crate::upsert_resolution::{Generation, TempIdMap, UniqueLookup};
 use crate::util::concat_bytes;
 
 // ---------------------------------------------------------------------------
@@ -297,7 +294,10 @@ pub async fn batch_lookup_unique_eids(
 ) -> Result<HashMap<(Entid, DataType), Entid>> {
     let mut prefixes: BTreeMap<Vec<u8>, (Entid, DataType)> = BTreeMap::new();
     for (attr_eid, value) in lookups {
-        prefixes.insert(unique_vae_prefix(*attr_eid, value), (*attr_eid, value.clone()));
+        prefixes.insert(
+            unique_vae_prefix(*attr_eid, value),
+            (*attr_eid, value.clone()),
+        );
     }
 
     let mut resolved: HashMap<(Entid, DataType), Entid> = HashMap::new();
@@ -409,173 +409,6 @@ pub async fn resolve_lookup_refs(
         });
     }
     Ok(result)
-}
-
-async fn resolve_temp_id_avs(
-    tempid_avs: &[(String, UniqueLookup)],
-    db: &slatedb::Db,
-) -> Result<TempIdMap> {
-    let mut unique_lookups: HashMap<UniqueLookup, Vec<String>> = HashMap::new();
-    for (tempid, lookup) in tempid_avs {
-        unique_lookups
-            .entry(lookup.clone())
-            .or_default()
-            .push(tempid.clone());
-    }
-
-    let lookups: Vec<UniqueLookup> = unique_lookups.keys().cloned().collect();
-    let resolved = batch_lookup_unique_eids(db, &lookups).await?;
-
-    let mut temp_id_map = HashMap::new();
-    for (lookup, tempids) in unique_lookups {
-        if let Some(&eid) = resolved.get(&lookup) {
-            for tempid in tempids {
-                temp_id_map.insert(tempid, eid);
-            }
-        }
-    }
-    Ok(temp_id_map)
-}
-
-fn record_resolutions(
-    resolved_tempids: &mut BTreeMap<String, Entid>,
-    temp_id_map: TempIdMap,
-) -> Result<()> {
-    let mut conflicts: BTreeMap<String, BTreeSet<Entid>> = BTreeMap::new();
-    for (tempid, entid) in temp_id_map {
-        if let Some(previous) = resolved_tempids.insert(tempid.clone(), entid) {
-            if previous != entid {
-                conflicts
-                    .entry(tempid)
-                    .or_default()
-                    .extend([previous, entid]);
-            }
-        }
-    }
-
-    if !conflicts.is_empty() {
-        return Err(anyhow::anyhow!("Conflicting upserts: {:?}", conflicts));
-    }
-    Ok(())
-}
-
-/// Per-tempid partition decision, computed once from the input datoms.
-///
-/// Encodes the only inference rule we use: a tempid asserting `:db/ident` is
-/// a schema entity and routes to `DB_PARTITION`; everything else routes to
-/// `USER_PARTITION`. Allocators downstream consult this map and never look at
-/// attribute content themselves.
-type TempIdPartitions = HashMap<String, u32>;
-
-fn infer_tempid_partitions(datoms: &[DatomWithTempids]) -> TempIdPartitions {
-    let mut out: TempIdPartitions = HashMap::new();
-    for d in datoms {
-        if let IdOrTempId::TempId(t) = &d.entity {
-            if d.op == DatomOp::Assert && d.attribute == kw!(:db/ident) {
-                // `:db/ident` always wins, regardless of ordering.
-                out.insert(t.clone(), DB_PARTITION);
-            } else {
-                out.entry(t.clone()).or_insert(USER_PARTITION);
-            }
-        }
-        if let ValueWithTempIds::TempRef(t) = &d.value {
-            out.entry(t.clone()).or_insert(USER_PARTITION);
-        }
-    }
-    out
-}
-
-fn allocation_partitions(
-    inferred: &TempIdPartitions,
-    tempid_labels: &BTreeMap<String, usize>,
-) -> Vec<u32> {
-    let label_count = tempid_labels.values().copied().max().map_or(0, |n| n + 1);
-    let mut partitions = vec![DEFAULT_PARTITION; label_count];
-    for (tempid, &label) in tempid_labels {
-        if let Some(&p) = inferred.get(tempid) {
-            partitions[label] = dominant_partition(partitions[label], p);
-        }
-    }
-    partitions
-}
-
-/// Resolve tempids using Mentat-style :db.unique/identity upsert generations.
-pub async fn resolve_tempids_with_upserts(
-    datoms: Vec<DatomWithTempids>,
-    schema: &Schema,
-    db: &slatedb::Db,
-    partition_map: &mut PartitionMap,
-) -> Result<Vec<Datom>> {
-    let inferred = infer_tempid_partitions(&datoms);
-    let (mut generation, inert_terms) = Generation::from(datoms, schema)?;
-    let mut resolved_tempids = BTreeMap::new();
-
-    while generation.can_evolve() {
-        let tempid_avs = generation.temp_id_avs();
-        let temp_id_map = resolve_temp_id_avs(&tempid_avs, db).await?;
-        record_resolutions(&mut resolved_tempids, temp_id_map.clone())?;
-        generation = generation.evolve_one_step(&temp_id_map, &resolved_tempids, schema)?;
-    }
-
-    generation.allocate_unresolved_upserts(schema)?;
-    // tempid -> label -> entid. Multiple tempid can map to the same entid. We first
-    // assign a label to each tempid group, then assign the partitions.
-    let tempid_labels = generation.temp_ids_in_allocations(schema)?;
-    let partitions = allocation_partitions(&inferred, &tempid_labels);
-
-    let mut allocated_tempids = HashMap::new();
-    let mut label_entids = Vec::with_capacity(partitions.len());
-    for partition in partitions {
-        label_entids.push(partition_map.allocate_entid(partition));
-    }
-    for (tempid, label) in tempid_labels {
-        allocated_tempids.insert(tempid, label_entids[label]);
-    }
-
-    let final_populations = generation.into_final_populations(&allocated_tempids)?;
-    let mut datoms = Vec::new();
-    datoms.extend(final_populations.upserted);
-    datoms.extend(final_populations.resolved);
-    datoms.extend(final_populations.allocated);
-    datoms.extend(inert_terms);
-    Ok(datoms)
-}
-
-/// Resolve tempids in DatomWithTempids to produce final Datoms.
-///
-/// Tempids with a `:db/ident` datom are allocated from DB_PARTITION;
-/// all others from USER_PARTITION.
-pub fn resolve_tempids(
-    datoms: &[DatomWithTempids],
-    partition_map: &mut PartitionMap,
-) -> Result<Vec<Datom>> {
-    let inferred = infer_tempid_partitions(datoms);
-
-    let mut tempid_map: HashMap<&str, i64> = HashMap::new();
-    for (tempid, partition) in &inferred {
-        let eid = partition_map.allocate_entid(*partition);
-        tempid_map.insert(tempid.as_str(), eid);
-    }
-
-    let mut resolved = Vec::with_capacity(datoms.len());
-    for d in datoms {
-        let entity = match &d.entity {
-            IdOrTempId::Id(id) => *id,
-            IdOrTempId::TempId(s) => *tempid_map.get(s.as_str()).unwrap(),
-        };
-        let value = match &d.value {
-            // TODO: get rid of the clone()
-            ValueWithTempIds::Data(data) => data.clone(),
-            ValueWithTempIds::TempRef(s) => DataType::Long(*tempid_map.get(s.as_str()).unwrap()),
-        };
-        resolved.push(Datom {
-            entity,
-            attribute: d.attribute.clone(),
-            value,
-            op: d.op,
-        });
-    }
-    Ok(resolved)
 }
 
 #[cfg(test)]
@@ -851,137 +684,5 @@ mod tests {
     #[should_panic(expected = "Delete/Erase not yet implemented")]
     fn test_expand_erase_panics() {
         expand_tx_ops(&[TxOp::Erase(EntityRef::Id(200))], &empty_schema()).unwrap();
-    }
-
-    // --- resolve_tempids tests ---
-
-    use crate::partition::{extract_counter, extract_partition, DB_PARTITION, USER_PARTITION};
-
-    #[test]
-    fn test_resolve_same_tempid_same_entid() {
-        let mut pm = PartitionMap::new();
-        let datoms = vec![
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("t1".to_string()),
-                attribute: kw!(:name),
-                value: ValueWithTempIds::Data(DataType::String("alice".to_string())),
-                op: DatomOp::Assert,
-            },
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("t1".to_string()),
-                attribute: kw!(:age),
-                value: ValueWithTempIds::Data(DataType::Long(30)),
-                op: DatomOp::Assert,
-            },
-        ];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        assert_eq!(resolved[0].entity, resolved[1].entity);
-    }
-
-    #[test]
-    fn test_resolve_different_tempids_different_entids() {
-        let mut pm = PartitionMap::new();
-        let datoms = vec![
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("t1".to_string()),
-                attribute: kw!(:name),
-                value: ValueWithTempIds::Data(DataType::String("alice".to_string())),
-                op: DatomOp::Assert,
-            },
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("t2".to_string()),
-                attribute: kw!(:name),
-                value: ValueWithTempIds::Data(DataType::String("bob".to_string())),
-                op: DatomOp::Assert,
-            },
-        ];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        assert_ne!(resolved[0].entity, resolved[1].entity);
-    }
-
-    #[test]
-    fn test_resolve_tempref_in_value() {
-        let mut pm = PartitionMap::new();
-        let datoms = vec![
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("alice".to_string()),
-                attribute: kw!(:name),
-                value: ValueWithTempIds::Data(DataType::String("alice".to_string())),
-                op: DatomOp::Assert,
-            },
-            DatomWithTempids {
-                entity: IdOrTempId::Id(999),
-                attribute: kw!(:follows),
-                value: ValueWithTempIds::TempRef("alice".to_string()),
-                op: DatomOp::Assert,
-            },
-        ];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        let alice_eid = resolved[0].entity;
-        assert_eq!(resolved[1].value, DataType::Long(alice_eid));
-    }
-
-    #[test]
-    fn test_resolve_db_ident_goes_to_db_partition() {
-        let mut pm = PartitionMap::new();
-        let datoms = vec![DatomWithTempids {
-            entity: IdOrTempId::TempId("schema-attr".to_string()),
-            attribute: kw!(:db/ident),
-            value: ValueWithTempIds::Data(DataType::Keyword(kw!(:my/attr))),
-            op: DatomOp::Assert,
-        }];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        assert_eq!(extract_partition(resolved[0].entity), DB_PARTITION);
-    }
-
-    #[test]
-    fn test_resolve_regular_tempid_goes_to_user_partition() {
-        let mut pm = PartitionMap::new();
-        let datoms = vec![DatomWithTempids {
-            entity: IdOrTempId::TempId("user-entity".to_string()),
-            attribute: kw!(:name),
-            value: ValueWithTempIds::Data(DataType::String("alice".to_string())),
-            op: DatomOp::Assert,
-        }];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        assert_eq!(extract_partition(resolved[0].entity), USER_PARTITION);
-    }
-
-    #[test]
-    fn test_resolve_id_passthrough() {
-        let mut pm = PartitionMap::new();
-        let datoms = vec![DatomWithTempids {
-            entity: IdOrTempId::Id(42),
-            attribute: kw!(:name),
-            value: ValueWithTempIds::Data(DataType::String("alice".to_string())),
-            op: DatomOp::Assert,
-        }];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        assert_eq!(resolved[0].entity, 42);
-        assert!(pm.is_empty(), "no allocation for explicit IDs");
-    }
-
-    #[test]
-    fn test_resolve_counter_advances() {
-        let mut pm = PartitionMap::from([(USER_PARTITION, 5_i64)]);
-        let datoms = vec![
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("a".to_string()),
-                attribute: kw!(:name),
-                value: ValueWithTempIds::Data(DataType::String("a".to_string())),
-                op: DatomOp::Assert,
-            },
-            DatomWithTempids {
-                entity: IdOrTempId::TempId("b".to_string()),
-                attribute: kw!(:name),
-                value: ValueWithTempIds::Data(DataType::String("b".to_string())),
-                op: DatomOp::Assert,
-            },
-        ];
-        let resolved = resolve_tempids(&datoms, &mut pm).unwrap();
-        let mut counters: Vec<i64> = resolved.iter().map(|d| extract_counter(d.entity)).collect();
-        counters.sort();
-        assert_eq!(counters, vec![5, 6]);
-        assert_eq!(pm[&USER_PARTITION], 7);
     }
 }
